Version 1.0.0 - Feature Complete and Fix Building
Fix an issue with builds failing with updated packages.

Marking as version 1.0.0 because there isn't any other features
I want to add. Will continue maintaining and fixing any bugs that
pop up in the future.

Feel free to submit any feature requests if there is something you'd
like and I will consider if I think it's in the scope of the plugin.